styling stuff, adding some privates, etc

+++ libswan/include/swan/common.h
@@ -7,7 +7,7 @@
 
 namespace Swan {
 
-static constexpr float TILE_SIZE = 32;
+static constexpr int TILE_SIZE = 32;
 static constexpr int TICK_RATE = 20;
 static constexpr int CHUNK_HEIGHT = 32;
 static constexpr int CHUNK_WIDTH = 32;
